Very basic collision. Started implementing live resolution checking.

+++ src/entity.cpp
@@ -1,7 +1,7 @@
 #include "entity.hpp"
 
 Entity::Entity(Graphics& Graphics, const std::string name,
-               const float passed_speed,
+               const double passed_speed,
                const Uint32 passed_single_frame_time_ms):
 Model(Graphics, name, passed_speed, passed_single_frame_time_ms)
 {
